implemeted rbt.

+++ Shader Viewer/Constants.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <glm/gtc/constants.hpp>
 
-static const double PI = glm::pi<double>();
-static const double EPS = glm::epsilon<double>();
-static const double EPS2 = EPS * EPS;
-static const double EPS3 = EPS2 * EPS;
+static const float PI = glm::pi<float>();
+static const float EPS = glm::epsilon<float>();
+static const float EPS2 = EPS * EPS;
+static const float EPS3 = EPS2 * EPS;
